+++ Games/Snake/main.cpp
@@ -9,13 +9,15 @@ int main(int argc, char *argv[])
     QSharedPointer<InterfaceReplica> ptr; // shared pointer to hold source replica
 
     QRemoteObjectNode repNode; // create remote object node
-    repNode.connectToNode(QUrl(QStringLiteral("local:interface"))); // connect with remote host node
+    repNode.connectToNode(QUrl(QStringLiteral("local:switch"))); // connect with remote host node
 
     ptr.reset(repNode.acquire<InterfaceReplica>()); // acquire replica of source from host node
 
-    ptr.data()->addCommand_slot("QQQ");
-    ptr.data()->rmCommand_slot("QQQ");
-    qio::qout << "QQQ" << endl;
+    while (true) {
 
+        ptr->addCommand_slot("QQQ");
+        ptr->rmCommand_slot("QQQ");
+        qio::qout << "QQQ" << endl;
+    }
     return a.exec();
 }
