--- UserInterface/StdCommands.h
@@ -1,25 +1,0 @@
-#ifndef STDCOMMANDS_H
-#define STDCOMMANDS_H
-
-#include "QString"
-#include "Command.h"
-
-namespace UI
-{
-
-class StdCommands
-{
-public:
-
-    StdCommands();
-
-private:
-
-    void help_request(const QString& pStr);
-    Command mHelp_request_cmd;
-
-};
-
-} // UI
-
-#endif // STDCOMMANDS_H
