+++ main.cpp
@@ -1,13 +1,11 @@
 #include <QCoreApplication>
 #include "UserInterface/Interface.h"
-#include "UIConfig.h"
 
 int main(int argc, char *argv[])
 {
     QCoreApplication a(argc, argv);
 
-    setUpUI();
-    UI::Run();
+    UI::Interface interface(&a);
 
     return a.exec();
 }
